+++ pyobjc-framework-XgridFoundation/PyObjCTest/test_xgfiledownload.py
@@ -1,9 +1,0 @@
-from PyObjCTools.TestSupport import TestCase
-import XgridFoundation
-
-
-class TestXGFileDownload(TestCase):
-    def testMethods(self):
-        self.assertArgIsBOOL(
-            XgridFoundation.XGFileDownload.setDestination_allowOverwrite_, 1
-        )