+++ pyobjc-framework-XgridFoundation/PyObjCTest/test_xggrid.py
@@ -1,7 +1,0 @@
-from PyObjCTools.TestSupport import TestCase
-import XgridFoundation
-
-
-class TestXGGrid(TestCase):
-    def testMethods(self):
-        self.assertResultIsBOOL(XgridFoundation.XGGrid.isDefault)